Adding new FlowData data structure
This is for easier managment of flow data after importing into a
standard struct. It inludes functions for gating, compensation, MEF
conversion, binning, and more.

This will help to standardize and streamline analysis workflows, though
plotting is still independent.

A key feature is interfacing with sampleMaps (eg from a text file) which
describe which samples were treated with which plasmids/small
molecules/etc.