Folders, updated binning, sampleMap editing
Added folder management, which causes automatically created graphs and
files to be placed into their respective directories (for different
repeats, for example).

Modified binning so that stats are not produced automatically. Updated
computeBinStats so it exports data in a mostly-matrix format, rather
than the complex struct like before

Added function for adding/editing sampleMaps on the fly