slice() can subsample numPoints and do controls
However, controls and bins cannot both be options since the controls are
not binned, so it will throw a warning and skip applying bins to the
data if slicing controls.

If 'equalize' and numPoints are both passed, then numPoints has
precedence. numPoints is automatically adjusted for the minimum number
of points per sample in the sliced dataset.

Also fixed an error where the default bin behavior of computeBinStats
was to do nothing, rather than iterate over every bin.