Playing around w/ comp functions
Trying to get it to have two modes: one for comp'ing all channels at
once and another where each channel is fit independently. However, I've
had to use the regression function for now because there is some issue
with the current code - to be fixed later, but this works fine for now.