Switched to all-MEFL transformations
Instead of using the unique MEF units for a given channel, convert to the equivalent (same bead) MEFL units instead.

This ends up giving the same product after two-color correction, but has the benefit of working even when the number of bead populations we are observing is higher than the number of quantified peaks for a given channel.

Made after discussion w/ Bre

Co-Authored-By: Breanna DiAndreth <brediandreth@users.noreply.github.com>