Update FlowData.m
Added additional options to threshGate and overall simplified a bit. In particular, the dataType input allows for thresholding on MEFL/comp/etc data rather than being limited to raw.

Also added sampleIDs input to enable selection of a subset of samples to use for generating threshold (useful for manually drawing when a given reporter is highly variable between samples and doesn't show up well when agglomerated).